Cleaned up for wider audience, added yaml example
