+++ .gitignore
@@ -2,3 +2,4 @@
 out.js
 .esbuild
 node_modules
+.DS_Store
