--- javascript.workaround/helloPlugin.js
@@ -1,18 +1,0 @@
-const helloPlugin = {
-  name: 'hello',
-  setup(build) {
-    console.log('Hello World')
-  },
-}
-exports.helloPlugin = helloPlugin;
-
-// const helloPluginFun = () => {
-//   return {
-//     name: 'hello',
-//     setup(build) {
-//       console.log('Hello World')
-//     },
-//   }
-// }
-//
-// exports.helloPlugin = helloPluginFun;
