+++ package.json
@@ -2,20 +2,20 @@
   "name": "plugin-missing-setup-function",
   "version": "0.0.1",
   "description": "",
-  "main": "app.js",
   "license": "ISC",
   "scripts": {
     "start": "cd javascript && serverless offline start --verbose",
     "start-ts": "cd typescript && serverless offline start --verbose",
-    "esbuild": "cd javascript && node esbuild.config.js",
-    "esbuild-ts": "cd typescript && ts-node esbuild.config.ts"
+    "start-yaml": "cd yaml && serverless offline start --verbose",
+    "build": "cd javascript && node esbuild.config.js",
+    "build-ts": "cd typescript && ts-node esbuild.config.ts",
+    "build-yaml": "cd yaml && node esbuild.config.js"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.44",
     "@types/node": "^18.7.3",
     "@types/serverless": "3.12.7",
     "esbuild": "^0.14.54",
-    "esbuild-plugin-copy": "^1.3.0",
     "serverless": "^3.21.0",
     "serverless-esbuild": "^1.32.5",
     "serverless-offline": "^8.8.1",
