--- typescript.workaround/app.txt
@@ -1,1 +1,0 @@
-You're a big cheese lovell, yes you are!
