--- typescript.workaround/helloPlugin.ts
@@ -1,14 +1,0 @@
-export const helloPlugin = {
-  name: 'hello',
-  setup(build: any) {
-    console.log('Hello World')
-  },
-}
-// export const helloPlugin = () => {
-//   return {
-//     name: 'hello',
-//     setup(build: any) {
-//       console.log('Hello World')
-//     },
-//   }
-// }
