--- typescript.workaround/tsconfig.json
@@ -1,33 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "baseUrl": ".",
-    "declaration": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "importHelpers": false,
-    "lib": ["esnext"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmit": true,
-    "rootDir": ".",
-    "sourceMap": true,
-    "target": "es2019",
-    "types": ["node", "jest"],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": false,
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "composite": true,
-    "incremental": true,
-    "noPropertyAccessFromIndexSignature": false,
-    "noImplicitOverride": false
-  },
-  "include": ["app.ts", "helloPlugin.ts"],
-  "exclude": ["node_modules"]
-}
