+++ typescript/helloPlugin.ts
@@ -1,14 +1,7 @@
+// module.exports = {
 export const helloPlugin = {
   name: 'hello',
-  setup(build: any) {
-    console.log('Hello World')
+  setup(config: any) {
+    console.log('Hello World: ')
   },
 }
-// export const helloPlugin = () => {
-//   return {
-//     name: 'hello',
-//     setup(build: any) {
-//       console.log('Hello World')
-//     },
-//   }
-// }
