+++ yaml/serverless.yml
@@ -1,0 +1,32 @@
+frameworkVersion: '3'
+service: hello-world
+plugins:
+  - serverless-esbuild
+  - serverless-offline
+custom:
+  esbuild:
+    entryPoints:
+      - app.js
+    platform: node
+    target: node16
+    bundle: true
+    outdir: .esbuild/.build
+    plugins:
+      - name: 'hello'
+# can't specify a function here!
+#        setup(config) {
+#          console.log('Hello World')
+#        }
+  serverless-offline:
+    host: localhost
+    httpPort: 3005
+provider:
+  name: aws
+  runtime: nodejs16.x
+functions:
+  hello-world:
+    handler: app.api
+    events:
+      - http:
+          path: hello
+          method: get
